requirements.txt
compas
pyside2
pyqt5
pyopengl
pyopengl-accelerate
